`zenml.artifacts.external_artifact.ExternalArtifact`
