+++ src/zenml/models/model_models.py
@@ -14,7 +14,7 @@
 """Model implementation to support Model WatchTower feature."""
 
 import re
-from typing import Any, Dict, List, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Type, Union
 from uuid import UUID
 
 from pydantic import BaseModel, Field, validator
@@ -30,6 +30,9 @@ from zenml.models.constants import STR_FIELD_MAX_LENGTH, TEXT_FIELD_MAX_LENGTH
 from zenml.models.filter_models import WorkspaceScopedFilterModel
 from zenml.models.model_base_model import ModelBaseModel
 from zenml.models.pipeline_run_models import PipelineRunResponseModel
+
+if TYPE_CHECKING:
+    from zenml.client import Client
 
 
 class ModelVersionBaseModel(BaseModel):
@@ -151,6 +154,11 @@ class ModelVersionResponseModel(
             for name, pr in self.pipeline_run_ids.items()
         }
 
+    def _import_client(self) -> Type["Client"]:
+        from zenml.client import Client
+
+        return Client
+
     def _get_linked_object(
         self,
         collection: Dict[str, Dict[str, UUID]],
@@ -174,7 +182,7 @@ class ModelVersionResponseModel(
         Raises:
             RuntimeError: If more than one object is found by given keys
         """
-        from zenml.client import Client
+        client = self._import_client()()
 
         search_pattern = re.compile(
             (pipeline_name or r"(.*)")
@@ -198,7 +206,7 @@ class ModelVersionResponseModel(
         name = names[0]
         if version is None:
             version = max(collection[name].keys())
-        return Client().get_artifact(collection[name][version])
+        return client.get_artifact(collection[name][version])
 
     def get_model_object(
         self,
