+++ src/zenml/steps/entrypoint_function_utils.py
@@ -27,11 +27,11 @@ from typing import (
 
 from pydantic import BaseConfig, ValidationError, create_model
 
+from zenml.artifacts.external_artifact import ExternalArtifact
 from zenml.constants import ENFORCE_TYPE_ANNOTATIONS
 from zenml.exceptions import StepInterfaceError
 from zenml.logger import get_logger
 from zenml.materializers.base_materializer import BaseMaterializer
-from zenml.steps.external_artifact import ExternalArtifact
 from zenml.steps.utils import (
     OutputSignature,
     parse_return_type_annotations,
