+++ src/zenml/steps/step_invocation.py
@@ -15,11 +15,11 @@
 from typing import TYPE_CHECKING, Any, Dict, Set
 
 if TYPE_CHECKING:
+    from zenml.artifacts.external_artifact import ExternalArtifact
     from zenml.config.step_configurations import StepConfiguration
     from zenml.new.pipelines.pipeline import Pipeline
     from zenml.steps import BaseStep
     from zenml.steps.entrypoint_function_utils import StepArtifact
-    from zenml.steps.external_artifact import ExternalArtifact
 
 
 class StepInvocation:
