+++ tests/integration/functional/zen_stores/test_zen_store.py
@@ -3172,27 +3172,32 @@ class TestModelVersionArtifactLinks:
                 ArtifactResponseModel,
             )
 
-            assert mv.model_objects["link2"]["1"].id == artifacts[1].id
+            assert (
+                mv.model_objects["pipeline::step::link2"]["1"].id
+                == artifacts[1].id
+            )
 
             assert (
                 mv.get_model_object("link2", "1")
-                == mv.model_objects["link2"]["1"]
+                == mv.model_objects["pipeline::step::link2"]["1"]
             )
             assert (
-                mv.get_deployment("link3", "1") == mv.deployments["link3"]["1"]
+                mv.get_deployment("link3", "1")
+                == mv.deployments["pipeline::step::link3"]["1"]
             )
 
             # check how versioned artifacts retrieved
             assert (
                 mv.get_artifact_object("link1", "1")
-                == mv.artifacts["link1"]["1"]
+                == mv.artifacts["pipeline::step::link1"]["1"]
             )
             assert (
                 mv.get_artifact_object("link1", "2")
-                == mv.artifacts["link1"]["2"]
+                == mv.artifacts["pipeline::step::link1"]["2"]
             )
             assert (
-                mv.get_artifact_object("link1") == mv.artifacts["link1"]["2"]
+                mv.get_artifact_object("link1")
+                == mv.artifacts["pipeline::step::link1"]["2"]
             )
